+++ server/src/network/session/session.rs
@@ -19,30 +19,19 @@ use crate::instance_manager::InstanceManager;
 use crate::level_manager::LevelManager;
 use crate::network::packets::login::{DeviceOS, Disconnect, PermissionLevel};
 use crate::network::packets::{
-    BroadcastPacket, ConnectedPacket, GameMode, MessageType, Packet,
-    PlayerListRemove, TextMessage,
+    ConnectedPacket, GameMode, MessageType, Packet, PlayerListRemove,
+    TextMessage,
 };
+use crate::network::raknet::BroadcastPacket;
 use crate::network::session::compound_collector::CompoundCollector;
 use crate::network::session::order_channel::OrderChannel;
 use crate::network::session::recovery_queue::RecoveryQueue;
 use crate::network::session::send_queue::SendQueue;
 use crate::network::Skin;
-use common::{bail, Serialize};
+use common::{bail, Serialize, Vector3f};
 use common::{error, VResult};
 
 use super::SessionManager;
-
-/// Tick interval of the internal session ticker.
-const INTERNAL_TICK_INTERVAL: Duration = Duration::from_millis(1000 / 20);
-/// Tick interval for session packet processing.
-const TICK_INTERVAL: Duration = Duration::from_millis(1000 / 20);
-/// Inactivity timeout.
-///
-/// Any sessions that do not respond within this specified timeout will be disconnect from the server.
-/// Timeouts can happen if a client's game crashed for example.
-/// They will stop responding to the server, but will not explicitly send a disconnect request.
-/// Hence, they have to be disconnected manually after the timeout passes.
-const SESSION_TIMEOUT: Duration = Duration::from_secs(5);
 
 const ORDER_CHANNEL_COUNT: usize = 5;
 
@@ -90,6 +79,12 @@ pub struct RaknetData {
 
 #[derive(Debug)]
 pub struct PlayerData {
+    /// Position of the player.
+    pub position: Vector3f,
+    /// Rotation of the player.
+    /// x and y components are general rotation.
+    /// z component is head yaw.
+    pub rotation: Vector3f,
     /// Game mode.
     pub game_mode: GameMode,
     /// General permission level.
@@ -157,6 +152,8 @@ impl Session {
 
             current_tick: AtomicU64::new(0),
             player: RwLock::new(PlayerData {
+                position: Vector3f::from([23.0, 23.0, 2.0]),
+                rotation: Vector3f::from([0.0; 3]),
                 runtime_id: RUNTIME_ID_COUNTER.fetch_add(1, Ordering::SeqCst),
                 game_mode: GameMode::Survival,
                 permission_level: PermissionLevel::Member,
@@ -182,71 +179,10 @@ impl Session {
             },
         });
 
-        // Start ticker
-        {
-            let session = session.clone();
-            tokio::spawn(async move {
-                let mut interval =
-                    tokio::time::interval(INTERNAL_TICK_INTERVAL);
-
-                while !session.active.is_cancelled() {
-                    match session.tick().await {
-                        Ok(_) => (),
-                        Err(e) => tracing::error!("{e}"),
-                    }
-                    interval.tick().await;
-                }
-
-                // Flush last acknowledgements before closing
-                match session.flush_acknowledgements().await {
-                    Ok(_) => (),
-                    Err(e) => {
-                        tracing::error!(
-                            "Failed to flush last acknowledgements before session close"
-                        );
-                    }
-                }
-
-                // Flush last packets before closing
-                match session.flush().await {
-                    Ok(_) => (),
-                    Err(e) => {
-                        tracing::error!(
-                            "Failed to flush last packets before session close"
-                        );
-                    }
-                }
-            });
-        }
-
-        // Start processor
-        {
-            let session = session.clone();
-            tokio::spawn(async move {
-                let mut broadcast_recv = session.broadcast.subscribe();
-                while !session.active.is_cancelled() {
-                    tokio::select! {
-                        pk = receiver.recv() => {
-                            if let Some(pk) = pk {
-                                match session.process_raw_packet(pk).await {
-                                    Ok(_) => (),
-                                    Err(e) => tracing::error!("{e}"),
-                                }
-                            }
-                        },
-                        pk = broadcast_recv.recv() => {
-                            if let Ok(pk) = pk {
-                                match session.process_broadcast(pk) {
-                                    Ok(_) => (),
-                                    Err(e) => tracing::error!("{e}"),
-                                }
-                            }
-                        }
-                    };
-                }
-            });
-        }
-
+        // Start processing jobs.
+        // These jobs run in separate tasks, therefore the session has to be cloned.
+        session.clone().start_ticker_job();
+        session.clone().start_packet_job(receiver);
         session
     }
 
@@ -267,6 +203,26 @@ impl Session {
         self.user_data.get().ok_or_else(|| {
             error!(NotInitialized, "User data has not been initialised yet")
         })
+    }
+
+    #[inline]
+    pub fn get_game_mode(&self) -> GameMode {
+        self.player.read().game_mode
+    }
+
+    #[inline]
+    pub fn get_position(&self) -> Vector3f {
+        self.player.read().position.clone()
+    }
+
+    #[inline]
+    pub fn get_rotation(&self) -> Vector3f {
+        self.player.read().rotation.clone()
+    }
+
+    #[inline]
+    pub fn get_permission_level(&self) -> PermissionLevel {
+        self.player.read().permission_level
     }
 
     /// Retrieves the identity of the client.
@@ -326,55 +282,6 @@ impl Session {
         self.raknet.guid
     }
 
-    /// Signals to the session that it needs to close.
-    pub fn flag_for_close(&self) {
-        self.initialized.store(false, Ordering::SeqCst);
-
-        if let Ok(display_name) = self.get_display_name() {
-            if let Ok(uuid) = self.get_uuid() {
-                tracing::info!("{display_name} has disconnected");
-                let _ = self.broadcast_others(TextMessage {
-                    message: format!("§e{display_name} has left the server."),
-                    message_type: MessageType::System,
-                    needs_translation: false,
-                    parameters: vec![],
-                    platform_chat_id: "".to_owned(),
-                    source_name: "".to_owned(),
-                    xuid: "".to_owned(),
-                });
-
-                let _ = self
-                    .broadcast_others(PlayerListRemove { entries: &[*uuid] });
-            }
-        }
-        self.active.cancel();
-    }
-
-    /// Sends a packet to all initialised sessions including self.
-    pub fn broadcast<P: ConnectedPacket + Serialize + Clone>(
-        &self,
-        pk: P,
-    ) -> VResult<()> {
-        self.broadcast.send(BroadcastPacket::new(pk, None)?)?;
-        Ok(())
-    }
-
-    /// Sends a packet to all initialised sessions other than self.
-    pub fn broadcast_others<P: ConnectedPacket + Serialize + Clone>(
-        &self,
-        pk: P,
-    ) -> VResult<()> {
-        self.broadcast.send(BroadcastPacket::new(
-            pk,
-            Some(
-                NonZeroU64::new(self.get_xuid()?)
-                    .ok_or_else(|| error!(NotInitialized, "XUID was 0"))?,
-            ),
-        )?)?;
-
-        Ok(())
-    }
-
     /// Kicks the session from the server, displaying the given menu.
     pub fn kick<S: AsRef<str>>(&self, message: S) -> VResult<()> {
         let disconnect_packet = Disconnect {
@@ -399,20 +306,5 @@ impl Session {
     #[inline]
     pub async fn cancelled(&self) {
         self.active.cancelled().await;
-    }
-
-    /// Performs tasks not related to packet processing
-    async fn tick(self: &Arc<Self>) -> VResult<()> {
-        let current_tick = self.current_tick.fetch_add(1, Ordering::SeqCst);
-
-        // Session has timed out
-        if Instant::now().duration_since(*self.raknet.last_update.read())
-            > SESSION_TIMEOUT
-        {
-            self.flag_for_close();
-        }
-
-        self.flush().await?;
-        Ok(())
     }
 }
